requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cassandra-driver==3.29.2
click==8.1.8
fastapi==0.115.6
geomet==0.2.1.post1
h11==0.14.0
idna==3.10
paho-mqtt==2.1.0
pydantic==2.10.4
pydantic_core==2.27.2
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
